requirements.txt
#作者：道长
admin==0.0.1
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
execnet==1.8.0
idna==2.7
importlib-metadata==2.0.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.4.0
mysqlclient==2.0.1
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.5.1
notebook==6.1.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
py==1.9.0
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==4.5.0
pytest-assume==2.4.2
pytest-emoji==0.2.0
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.20.0
Send2Trash==1.5.0
six==1.15.0
tblib==1.7.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.2.0
